moves model object logic into partial class files
