+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Parameters/SpecklePropertyGroupWrapper.ModelObjects.cs
@@ -1,0 +1,32 @@
+#if RHINO8_OR_GREATER
+using Grasshopper.Rhinoceros;
+using Grasshopper.Kernel.Types;
+
+namespace Speckle.Connectors.GrasshopperShared.Parameters;
+
+/// <summary>
+/// The Speckle Property Group Goo is a flat dictionary of (speckle property path, speckle property).
+/// The speckle property path is the concatenated string of all original flattened keys with the property delimiter
+/// </summary>
+public partial class SpecklePropertyGroupGoo : GH_Goo<Dictionary<string, SpecklePropertyGoo>>, ISpeckleGoo
+{
+  private bool HandleModelObjects(object source)
+  {
+    if (source is ModelUserText userText)
+    {
+      Dictionary<string, SpecklePropertyGoo> dictionary = new();
+      foreach (KeyValuePair<string, string> entry in userText)
+      {
+        string key = entry.Key;
+        SpecklePropertyGoo value = new() { Path = key, Value = entry.Value };
+        dictionary.Add(key, value);
+      }
+
+      Value = dictionary;
+      return true;
+    }
+
+    return false;
+  }
+}
+#endif
